requirements.txt
numpy
pandas
streamlit
pickle-mixin
google-generativeai